LayoutTests/imported/w3c/web-platform-tests/html/interaction/focus/the-autofocus-attribute/no-cross-origin-autofocus-expected.txt

CONSOLE MESSAGE: Blocked autofocusing on a form control in a cross-origin subframe.
Autofocus shouldn't work in cross-origin iframe.



PASS Autofocus shouldn't work in cross-origin iframe
